refactor(Project): switch to POJOs
