+++ src/Components/ProjectForm.tsx
@@ -1,103 +1,17 @@
 import { ChangeEvent, Dispatch, FormEvent, SetStateAction } from 'react';
-import { DateString, UUIDString } from '../lib/types';
-import generateUniqueId from '../lib/uniqueId';
+import { UUIDString } from '../lib/uniqueId';
 import Input from './Input';
 import TextArea from './TextArea';
+import { formatDate } from '../lib/dateUtils';
 
-class ProjectDetails {
+interface ProjectDetails {
   id: UUIDString;
   name: string;
   techUsed: string;
   description: string;
   isVisible: boolean;
-  #startDate: Date = new Date();
-  #endDate: Date = new Date();
-  [key: string]: string | boolean | string[] | (() => ProjectDetails);
-
-  constructor({
-    id = generateUniqueId(),
-    name = '',
-    techUsed = '',
-    description = '',
-    startDate = new Date(),
-    endDate = new Date(),
-    isVisible = true,
-  } = {}) {
-    this.id = id;
-    this.name = name;
-    this.techUsed = techUsed;
-    this.description = description;
-    this.startDate = startDate;
-    this.endDate = endDate;
-    this.isVisible = isVisible;
-  }
-
-  set startDate(date: Date | DateString) {
-    this.#startDate = date instanceof Date ? date : this.#parseDate(date);
-  }
-
-  get startDate(): DateString {
-    return this.#formatDate(this.#startDate);
-  }
-
-  set endDate(date: Date | DateString) {
-    this.#endDate = date instanceof Date ? date : this.#parseDate(date);
-  }
-
-  get endDate(): DateString {
-    return this.#formatDate(this.#endDate);
-  }
-
-  get duration() {
-    const start = `${ProjectDetails.shortenDate(this.#startDate)}`;
-    const endMonth = this.#endDate.getMonth();
-    const endYear = this.#endDate.getFullYear();
-
-    const presentDate = new Date();
-    const thisMonth = presentDate.getMonth();
-    const thisYear = presentDate.getFullYear();
-
-    const end =
-      endMonth === thisMonth && endYear === thisYear
-        ? 'Present'
-        : `${ProjectDetails.shortenDate(this.#endDate)}`;
-
-    return [start, end];
-  }
-
-  static shortenDate(date: Date) {
-    const userLocale = navigator.languages[0];
-    const formatter = new Intl.DateTimeFormat(userLocale, {
-      month: 'short',
-      year: 'numeric',
-    });
-
-    const [month, year] = formatter.format(date).split(' ');
-    return `${month}. ${year}`;
-  }
-
-  clone(): ProjectDetails {
-    return new ProjectDetails({
-      id: this.id,
-      name: this.name,
-      techUsed: this.techUsed,
-      description: this.description,
-      isVisible: this.isVisible,
-      startDate: this.#startDate,
-      endDate: this.#endDate,
-    });
-  }
-
-  #formatDate(date: Date): DateString {
-    const year = date.getFullYear();
-    const month = (date.getMonth() + 1).toString().padStart(2, '0'); // 0 indexed month
-    const day = date.getDate().toString().padStart(2, '0');
-    return `${year}-${month}-${day}`;
-  }
-
-  #parseDate(dateString: DateString): Date {
-    return dateString.length === 0 ? new Date() : new Date(dateString);
-  }
+  startDate: Date;
+  endDate: Date;
 }
 
 interface ProjectFormProps {
@@ -124,13 +38,13 @@ function ProjectForm({
     const field = e.target.name;
     const value = e.target.value;
     const originalDetails = projectDetails.at(detailIndex)!;
-    const newDetails = originalDetails.clone();
-    newDetails[field] = value;
+    const newDetails = { ...originalDetails, [field]: value };
 
     const newProjectDetails = projectDetails.map((details) => {
       if (details.id === id) return newDetails;
       return details;
     });
+
     setProjectDetails(newProjectDetails);
   }
 
@@ -169,14 +83,15 @@ function ProjectForm({
         label='Start Date: '
         type='Date'
         name='startDate'
-        value={startDate}
+        value={formatDate(startDate)}
         onChange={handleFieldChange}
       />
+
       <Input
         label='End Date: '
         type='Date'
         name='endDate'
-        value={endDate}
+        value={formatDate(endDate)}
         onChange={handleFieldChange}
       />
       <button className='btn-save' type='submit'>
@@ -186,5 +101,5 @@ function ProjectForm({
   );
 }
 
-export { ProjectDetails };
+export type { ProjectDetails };
 export default ProjectForm;
